gotohelm: add partial support for `resource.Quantity`
This commit adds partial support for Kubernetes' `resource.Quantity` type into
gotohelm. The supported methods are:
- `resource.MustParse`
- `resource.(*Quantity).Value`
- `resource.(*Quantity).String`
- `resource.(*Quantity).DeepCopy`

It is also possible to include a `resource.Quantity` type on a Values struct
and interact with it as you would any other `resource.Quantity` provided that
the value is a numeric or a valid `resource.Quantity` string.

+++ pkg/gotohelm/testdata/src/example/sprig/sprig.rewritten.go
@@ -78,8 +78,8 @@ func trim() []string {
 	}
 }
 
-func minFunc() []int {
-	return []int{
+func minFunc() []int64 {
+	return []int64{
 		// spig Min function does not allow to not pass parameters. Error returned from helm template:
 		// wrong number of args for min: want at least 1 got 0
 		// helmette.Min(),
